create and delete functions for ctl
